Update mobile-services-html-get-started-users.md
The last note states that you should change `facebook` for `windowsazureactivedirectory`. 
I'm not able to make authentication work by changing it to `windowsazureactivedirectory`. It gives me a **401** with the message "Error: Authenticating with 'windowsazureactivedirectory' is not supported.".

If I change it to `aad` (deduced from the url given to me by azure...) instead, it works like a charm.