Fixing duplicate WACOM.NOTE
If there are two [WACOM.NOTE]s in a row then the second one is ignored and the actual text [WACOM.NOTE] gets output, which appears to be an error in the backend Markdown to HTML generator. http://azure.microsoft.com/en-us/documentation/articles/virtual-machines-windows-tutorial/

Also adding some additional prose to add clarity to the VHD section.