Update command-line-tools.md
The vm create command was wrongly stating that its --blob-url option specifies the blob from which the system disk is created. There is actually no option which does that. The --blob-url option actually does the opposite: rather than specifying the source blob, it specifies where  to create the target blob.