Update web-sites-java-custom-upload.md
In an attempt to improve documentation/examples such that they will work with Azure Pack websites as well we need to change the references to D:\home to %HOME%.  In Azure Pack the drive location is C:\home.  In the event there is a need to find the feature owner to approve, I'm the PM for both the Java feature and Azure Pack Websites and I approve of myself making these changes. :-)
Chris Compy