Add valid values for --vm-size on vm create
I know that they are many values here, but I think is useful for people who wants to automate the process. I have been building an app that let's the user select the size of the machine and since I didn't see the values values in the first place I thought it wasn't possible.