The API v1 documentation does not need to change.
The API v1 documentation does not need to change. The next version documentation is up to date as well, and posted to https://azure.microsoft.com/en-us/documentation/articles/machine-learning-recommendation-api-documentation/.
Will need to discuss with the team if we should keep this page published or not. For now keep it.