Updated to prep for html tag removal.
Removed the <mark> tags and changed text so that there is no mention of highlighted text. This topic is now ready to have the html tags removed from the doc. I believe from the notes, there are only <pre> tags that need to be converted (removed from the markdown). Please do not modify html that is actually part of the tutorial.