Updating disk drive requirements
This change needs to be published as soon as possible. We have been getting a lot of customer escalations because customers ship the wrong disk type for Import/Export service. Adding an explicit note to clarify that external HDD are not supported. 

Please help publish this as soon as possible.