Update automation-dsc-cd-chocolatey.md
I've excised the heavy "I" usage because, once you move to a new position, someone else will have to maintain it.... although you'll be the owner and have your face and name up there, it's hard to transfer the "I" tone to a new human who isn't "I". :-) Everything else looks great. 

If you accept the changes -- or even if you reject them -- we can accept the PR.