Update sql-database-understanding-dtus.md
DTU stands for database throughput unit not database transaction unit.  While DTU is a representation of the overall ability of the database to process transactions it's not an exact correlation.  For consistency with the other pages on the Azure product pages/documentation we should be using database *throughput* throughout (see for example: https://azure.microsoft.com/en-us/pricing/details/sql-database/) .  I've also added an metaphor of horsepower to help people understand the notion of DTU in terms that most people are at least familiar with.