Attempted to clarify what Tenant ID means
As a user who isn't knowledgeable about Active Directory, I didn't find the Tenant ID description to be helpful. In particular, "contoso.com" was very misleading to me because Tenant IDs seem to usually be Guids. Also it turns out that determining your Tenant ID is very easy with Azure PowerShell, but it took my a while to figure out.