Make it easier to integrate with Azure web sites
It adds a reference link on how to configure client certificates in Azure Websites. I spent the full day getting this to work due to lack of knowledge about how to do it on websites (misleading that we can remote into IIS Manager and activate it there, even though that is the wrong way to do it). Until I finally found the link referenced and got it to work. If this link would have been part of the article, I would have finished configuring in 1 hour and not 8.