Retrieval of connection string from web.config/app.config is wrong
The code for retrieving the storage accounts connection string was wrong. The connection string should either be created under configuration > connectionStrings or the retrieval should point to appSettings. As there are multiple documents already referring to the string being stored under appSettings I propose to change the retrieval code here.