Search: Removing link to GitHub sample from .NET How-to
The .NET How-to article is going to be replaced soon with a new Getting
Started with .NET article series. The sample code on GitHub is being updated
for the new series and will no longer match what's in this article. However,
this article has its own full copy of the source already, so it is self-
contained.