Changed names of values to reflect those on Azure Portal
The name of the route properties are outdated and do not currently reflect the names that are on the Azure Portal, causing confusion when deploying UDR routes on the portal or even when trying to do using Azure PowerShell. I have updated the names to correctly match those of the portal and added a small section to specify the differences in names when using Azure PowerShell.