M653221 Fix "below" to "above"
@VybavaRamadoss
This sentence "More details in the access control section below"  appears twice in this article. The second one appear below the section on "access control". Also, the link does not work in both sentences, because there is not header "#access-control-for-a-tenant-resource-provider".
Could you please help to review and fix the link as well?
@SunnyDeng