Guide users to not install R versions of PowerCLI
I had an issue at a customer where we installed the latest R version of PowerCLI on our processing server which caused a significant amount of troubleshooting for a week.  We lost a lot of time.  We couldn't uninstall the R3 version and just install "RTM" as when R3 is installed, the code is stuck in there even when installed 6.0.  We had to rebuild our processing server from scratch.  When you look at the download page, I believe you are compelled to want to get the latest version, which is the issue we got stuck on.  The edit is to try and avoid other pros from making this mistake.