M659382: Possible copy/paste mistake - "you are planning to add to the cluster"
@ChackDan,
It seems that in line 150 the phrase "you are planning to add to the cluster" does not make sense because the preceding step is "Remote desktop (RDP) into the VM/machine that you want to remove from the cluster". Also, if you take a look at line 139, the same sentence is included and it seems to be correct since it belongs to paragraph introduced by "Add nodes to your cluster" .
Please, check the proposed deletion and sign off if you agree.
Many thanks in advance. 
CC: @v-jacard