Ensure WindowsFeature Web-IP-Security Is On
We have found that unlocking the section is not enough and that the IP security feature is not switched on by default. Caused a bit of a delay to getting IP sec up and running.

Got the above script from Satya Ramani at azure support and feel it would have been really helpful if it had been here, which was the first place we looked.