For clarity - NIC requires existing VNet
From:
The NIC must be connected to a subnet within an existing Azure Virtual Network (VNet) in the same location and [subscription](../azure-glossary-cloud-terminology.md#subscription) as the NIC. 

To:
The NIC must be connected to a subnet within an existing Azure Virtual Network (VNet). The three components: NIC, subnet, and VNet, must all exist in the same region and [subscription](../azure-glossary-cloud-terminology.md#subscription).