Update automation-hybrid-runbook-worker.md
There was a lot of false information about SMA. I adjusted for a more fair comparison.
I deleted the statement where it was implied that Azure Automation was preferred because it could manage both Azure and On-prem where SMA supposedly could not. This is false statement as SMA Runbook workers can interface with Azure just as easily.