Update azure-stack-key-features.md
Internal use is going to be more prevalent than service providers amongst the readers so change the emphasis.

Portal and PowerShell are co-equal paths for interacting with AS.

"consistent with Azure Portal" implies "different".  It isn't different, it is an instance of the Azure Portal running on your servers.

Services and applications - the examples are services => applications so the setup words should follow the same order.