Updating "Important" note
Updated the "Important" Note that we have in our page to be more clear. Unchecking Prefer 32-bit, only applies to Executable projects, for rest of the project types there are different ways but didn't wanted to list them all. Just added a sentence: Ensure that your host process is 64 bit.