Update virtual-machines-windows-sql-high-availability-dr.md
Added a 3rd option for FCI - using Windows Server 2016 Storage Spaces Direct.  We built this for a customer based on feedback from the product group and was able to see that SQL Server supported using this shared storage, SQL Server operated as expected and failover functioned as expected (when we turned off one of the VMs).