Update virtual-machine-scale-sets-deploy-app.md
Added link to the Azure document that provides further information on the process to roll out an OS upgrade across a VMSS. I also removed the link to the vmssupgrade tool since it's mentioned on the linked doc and is a bit redundant. Can add it back in though if you want to keep it on this page.