Fix: redisCacheLocation is no longer used by the referenced template
See change history of https://github.com/Azure/azure-quickstart-templates/blob/master/101-redis-cache/azuredeploy.json - now it deploys to the resource group location, in line with some of their recommended best practices: "Do not use a parameter to specify the location. Use the location property of the resourceGroup instead. By using the resourceGroup().location expression for all your resources, the resources in the template will automatically be deployed in the same location as the resource group."