Replaced command for 'activating AD DS powershell'
Its not a good idea to get people to run 'Install-WindowsFeature –Name AD-Domain-Services –IncludeManagementTools' just to install the management tools... this will install AD DS on that server! And since this is the AAD Connect server, that's not ideal! I've replaced it instead with `Add-WindowsFeature RSAT-AD-PowerShell` which instead just does exactly as it says on the tin.