Fixed incorrect wording of bullet #2
Wording for bullet 2 in "Why you may need an internal load balancer" pointed to connectivity to VMs within the same cloud service, but referred to figure 2 which indicates connectivity to "Front-End servers in a different cloud service".  Updated the bullet text to indicate the "different cloud service" - but also to indicate the potential of no cloud service for ARM (IaaS v2) VMs.