Update wording (visible -> invisible) of comment
Updated the wording of the comment to avoid confusion line 164.  The existing comment "Make the message visible" is inaccurate and indicates the message will again be visible in the queue after the call.  The line of code in fact extends the period to which the message will be invisible to other clients for an additional 60 seconds.