Update documentdb-sdk-dotnet-core.md
@mimig1 @arramac
Not sure for some reason why my previous PR for updating the .NET Core Release Notes is not showing up, so here is the new one.

Mimi, for the following line, this will be applicable to all future releases of .NET Core SDK, so I've added it under the Release Notes. Feel free to highlight it appropriately if needed.
The DocumentDB .NET Core SDK has feature parity with the latest version of the [DocumentDB .NET SDK](documentdb-sdk-dotnet.md).

I've added the Retirement date for the preview bits to be Dec 31.