Correct spelling of pre-requisite to prerequisite
This change corrects the spelling of pre-requisite in the url link title and does NOT change the spelling in the reference to the markdown file.  This also makes the url link title match the spelling used in the title of the referenced markdown article.  
Note: A partner noticed and mentioned this inconsistency to me.