Replaced command for 'activating AD DS powershell'
As previously submitted here (https://github.com/Microsoft/azure-docs/pull/156), but since that file has been shelled to just a redirect, I've copied the pull request to the new file...

Its not a good idea to get people to run 'Install-WindowsFeature –Name AD-Domain-Services –IncludeManagementTools' just to install the management tools... this will install AD DS on that server! And since this is the AAD Connect server, that's not ideal! I've replaced it instead with Add-WindowsFeature RSAT-AD-PowerShell which instead just does exactly as it says on the tin.