Update doc to reflect Azure AD in new portal
Azure AD is now available in the new portal at portal.azure.com. We should revise the document to also show how to set up Azure AD auth using the new portal, as there are a few pieces that are different.

I've started some changes, but there are major changes that need to be made. Navigating to directories is different (done by clicking your user in the top right of the screen and going to the new directory). Apps are under App registrations now. A few others, too.