Update documentdb-resources.md
There was a sidenote comment around "DocumentDB account" vs "database account" and I clarified that these are used interchangeably and mean the same thing. Since we use the term "database account" more commonly in the documentation, I changed the text in one place. I don't think it's worth to get the picture updated.
@anhoh @mimig1 FYI