Update registry keys for Visual Studio
The registry keys in this document have not worked a few cases we've received this topic. These the registry keys that we are using to resolve VS 2015 connectivity with Azure Government. Also, please note that I recommend that if the changes are accepted, then I added Visual Studio 2015 to the title because the extra set of registry keys seem to be specific to VS2015. Also, you may want to place the default warning we put on all articles that recommend a registry key change. Also, if an account failed to add, fixing the registry keys will require the user to delete the account and re-add the account using Storage Explorer/Cloud Explorer.