Possible double-quotes in $packageUri
While using this article to integrate Existing Web Apps to VNETs in Azure Government, we have noticed that the Get-AzureRmVpnClientPackage cmdlet was returning a URL with literal double-quotes at the beginning and the end of the URL string. This caused the New-AzureRmResource to save the gateway configuration with a blank package Uri. Hence the change proposed tries to address identifying if the URL returned contains a literal double-quote at the beginning and the end of the URL and strips these double-quotes off before calling New-AzureRmResource.