Update service-bus-dotnet-hybrid-app-using-service-bus-relay.md
I've added "primary" and "secondary" before SAS key to make it clear, which keys you are supposed to copy in. Whilst following along the tutorial I used the primary key in the web application too and ran into an exception. 
Since there are 2 keys in the Azure portal, it should be clear to the reader, which one to use.