Fix NAT rule name conflicts
At first, `ssh1` and `ssh2` were created as NAT rules for load balancer, but then, `rdp1` and `rdp2` were used to create NICs, which caused the conflicts.

Considering that it will create two windows VMs for the load balancer, it makes sense to use `rdp` rather than `ssh` for a quick example. Of course, if you think `ssh` is better, I'd glad to change it back.