Provide example use of private Azure Blob Storage
Provided an example of using private Azure Blob Storage for scripts/resource in a storage account that exists in the same subscription but a different resource group to store the scripts and other resources. Additionally described the rationale and method required for using listKeys instead of manually providing the storage account key to access resources stored in a private Azure Blob.