Remove SAML 2.0 as option
Using the usual Azure AD custom app integration path with SAML 2.0 does not work with AppSource because it requires admin intervention.  All AppSource apps need to be user-consentable.  User consent with SAML 2.0 is possible but we do not have published guidance today on how to do it.  Removing SAML 2.0 as an option until we have that guidance published.