Added Container Variable and Note
Errors occur later in the commands because the variable for the
container wasn't created. Also, an error occurs if the context is not
set. The error indicates Microsoft.Insights needs to be registered, but
the error persists even after registering Microsoft.Insights. This
happens specifically when the user does not set the context and has
multiple accounts.