Minor Changes
Container variable was missing in the documentation. There are commands
that reference the $container variable, but it did not exist and causes
an error to the end user.

Added a note to highlight the importance of running the
Set-AzureRMContext, which also generates an obscure error to register
Microsoft.Insights if the user does not set the context. The error will
still persist even if the user registers Microsoft.Insights.