Made a few edits on review.
One thing i noticed is that we should probably convert the code sample to a blob trigger example instead of an HTTP trigger.  I think that the blob trigger is the easiest "100" level thing to understand for customers who just want us to do some encoding on a video. We should be referencing the 100 level Encoding Function sample