Fixed description of how template expressions work
Removed text indicating that [ ] expressions can occur anywhere within a string literal.  In fact, the first character of the string must be the opening bracket.   In reality, that must also be followed by the name of a built-in function, but that would be confusing at this point in the doc.   Also reworded the paragraph for clarify.

There should ultimately be text added to clarify that template expressions are only evaluated in literals occurring on the "value" side of key/value entries.  For example, you can't say
 { ...    "[variables('myKeyName')]" : "some value" 
because ARM template processing doesn't look for expressions on the left-hand-side.

selliott@microsoft.com