Update virtual-machines-windows-extensions-customscript.md
Hello. I cannot find anything that indicates how often the commands contained in the custom script extension are executed. Is it when the machine deployed? Is it ran each time the VM is rebooted? Is it run on a schedule? Can I force the extension to execute the commands it contains anytime I want? etc