Update resource-manager-tag-resources.md
Made a minor change to the wording of step 4: "To view all the resources with a tag value". The existing phrase "filter by Tags" is confusing, as there is a filter drop down next to the text box with the choices "by category" and "by name". But the instructions are not to use the filter drop-down, but to enter the word "tags" into the text box.