Fixing few Critical Bugs in the documentation after testing
Changing the name in az disk list command to be consistent with "az disk create". 
Also to create a VM with Managed Disks you need to provide the ID. URI works only for Unmanaged Disks. Updating the az disk list and "az vm create" commands to reflect this in the "Azure CLI 2.0 - Azure Managed Disks" section. 
Also, the quotes in the value for --query in "az disk list" command are not correct. Azure CLI provides result only when the single quotes are not provided.