Update Default Tags to match ARM behaviour
If you're using Azure Resource Manager, the Default Tags have subtly different names - as https://docs.microsoft.com/en-us/rest/api/network/create-or-update-a-network-security-rule shows, the canonical names are `VirtualNetwork`, `AzureLoadBalancer` and `Internet`. Due to ARM apparently being case insensitive, it will accept `INTERNET` as documented in this page, but not `VIRTUAL_NETWORK` or `AZURE_LOADBALANCER` due to the underscores.

This caused me to run into a mystifying error when I tried to create an NSG using these default tags, and I'm not the first to be confused by this as you can see from https://github.com/Azure/azure-powershell/issues/391#issuecomment-188118549

Since the top of this page recommends using ARM, the page should at the very least mention the Default Tag names required by ARM. (Arguably, it should put them first, and mention parenthetically that if you're working in the world of Classic, the older SHOUTING_CAPS_WITH_UNDERSCORES names should be used instead. But I wanted to propose as minimal a change as possible.)