Update howto-connection-string.md
It's not immediately clear that you need to use userID@servername as the format for any user account that authenticates to this server.  I spent two days trying to troubleshoot why I couldn't connect to the server with my webapp while I could use MySQL Workbench to connect with the user account.  Specifying userID@servername is what solved this problem for me.