Flag --disable-nodejs-remote-modules was removed from build script
A recent change removed this flag, so you cannot pass it to the Windows build script anymore. Node.js remote modules are now disabled by default.

Also, added pkg-config to the list of dependencies to `apt-get`. This package was also added to `doc/devbox_setup.md` in the `Azure/iot-edge` repo.