Bandwidth can limit performance improvements
If the queries in the example are ran it can be that slow network connections is saturated. The effect is that performance improvements are not seen as the receiving machine cannot get the results fast enough. This is particularly the case if on a slow DSL line like me :(. Added a note to make this clear.