Azure Backup documentation edit
My name is Vikranth, from the MSDN Azure forums team.

I would like to suggest a few edits/corrections in the following documentation. 
Monitor and manage Azure recovery services vaults and servers for Windows machines.

In the highlighted steps/statement in the documentation “You can also set Exclusion Settings from this page in the wizard. If you want to exclude files or file types read the procedure for adding exclusion settings.” 
The exclusion settings, hyperlink is not pointed to the correct section. 

It is now pointing to https://docs.microsoft.com/en-us/azure/backup/backup-azure-manage-windows-server#exclusion-settings

But, it should point to this section: https://docs.microsoft.com/en-us/azure/backup/backup-azure-manage-windows-server#manage-exclusion-settings.

Requesting you to correct this section in the documentation.