Update vpn-gateway-howto-site-to-site-classic-portal.md
The original document has customers logging in to ARM in PowerShell then logging in to ASM. Logging into ARM in this context is superfluous and has caused confusion to our customers, especially since an additional step selecting the right subscription in ASM was omitted. 
I propose removing the PowerShell ARM login commands and replacing them with the appropriate ASM ones.