Add instructions on how to remove a backup
Would be good to add instructions on how to remove a backup since you cannot do it via the portal, and if a recovery vault has a backup in it, it can't be deleted. Thus if you try out Azure Backup, it is hard to undo (and thus you will be charged for it).