Add step for changing default subscription
I had to search on SO to find out how to change my default subscription since I've got several. It seems like this would be a useful place to tell people how to select the right subscription since that may be part of the workflow. I do **not** know if changing the default from the CLI like this will have unintended consequences for the user elsewhere in the portal.