Missing policy creation for Facebook auth
This help doc on Facebook auth should include the steps that are required to configure a policy; otherwise, auth isn't configured properly and it doesn't work.  Also, when following the current set of instructions, it's not necessarily apparent that these steps are missing.  I also noticed that another dev encountered this same issue and added a comment to this page.  This same issue exists across all of the guides for configuring social auth.