Update storage-use-azcopy-linux.md
Updated dotnet to 1.0.4 as the latest version recommended on the https://www.microsoft.com/net/core#linuxubuntu page.  Also added paragraph showing lsb_release -a command to help user determine which release of Ubuntu they are using so that they can navigate to the right section.  The dotnet and azcopy sections could still do with a little more tidying.