Change verbiage on resource groups
Resource groups is a bit of cognitive overhead for developers and injects a paralysis point where a developer might just stop the tutorial because they don't know what to do here. This PR tries to simplify the concept, clear up pricing and offer a suggestion about how to handle this situation.

Note the part on "does not directly affect price". I'm not entirely sure if this is correct. I need someone with better pricing knowledge to sign off on that verbiage.