Removing unnecessary styles
Removing styles that are unnecessary for modern browsers with Bootstrap. We could probably do away with these styles altogether and just use Bootstrap containers, but this is ok too.

I do **NOT** know if this breaks compatibility in older browsers (IE 11 and lower). It's been a long time since I had to consider that someone might be on an old browser version.