Case changes/Clarfiying/Adding Note
I made all references to HCs initial capped since it's a proper noun. I also reworded a few sentences to make them clearer. Finally, I added a note about Windows Server 2008 R2 not supporting Web Sockets.

One thing I didn't note is that connectivity to Service Bus gateways is also required. I'm not sure if you wanted to add that info to this doc or not, but HCs do go through one or more (up to 128) Service Bus gateways, and if outbound connectivity from the HCM machine is blocked to those URLs, HCs will fail. Not a common scenario, but we've seen it with one customer already.