replaced cosmosdb_name with cosmosdb-name
when the name is `cosmosdb_name` some folks may try the command by using this value. The underscore is not a valid character for a database account, so I propose we change this to a dash. 

See below for details ...

```
$ az cosmosdb create --name my_cosmosdb_name --resource-group myResourceGroup --kind MongoDB 
The character '_' at index 2 is not allowed in the DatabaseAccount name
ActivityId: 01fa2151-605d-11e7-9bf0-784f4394b80b
```