making copying the keys clearer
When we get the keys, we tell the user to copy the primaryMasterKey from the output json in terminal.

az cosmosdb list-keys --name my-cosmosdb-name --resource-group myResourceGroup 

Instead we could tell the user to run this command which outputs just the primaryMasterKey, which makes it easier to copy and paste for the reader.

az cosmosdb list-keys --name my-cosmosdb-name --resource-group myResourceGroup --query "primaryMasterKey"

we just told the user to copy the key. but then we tell them to copy another snippet of code and paste it. this means key is not retained in the clipboard's most recent value. Instead, I propose we switch the order of these to make it easier for the reader.