Emphasizing that this is for preview
The tools don't say that it requires a preview version of VS, nor does the error in the installer. 
It's mentioned in one line at the top. I've found other people who have run into the same issue when installing it, so hopefully this prevents others from having the same problem.