Update certificate pinning guidance (from SDL)
The current pinning guidance was incorrect, as it did not perform basic validation, and did not check to ensure the algorithm was expected. I re-wrote the guidance to match our general SDL guidance on pinning, though this could be further expanded to talk about mechanisms for updating and having sensible fallbacks.