Correct ARM Template for Azure file share example
The azuredeploy.json file had invalid resources node and also was missing the port definitions so that a user could access the web app to interact with the Azure file share.

I added some additional context around obtaining the ip address of the running container so that a user could load the web app into their browser and write content to the Azure file share. I also added guidance on an easy method to retrieve and inspect the files written by the web app.