Update storage-faq-for-disks.md
**How can I encrypt my existing disks that were created before June 10, 2017?**

As of June 10, 2017, new data written to the underlying storage accounts of managed disks is automatically encrypted. We are also planning to encrypt existing data, and the encryption will happen asynchronously in the background. If you must encrypt existing data now, create a copy of your disk. New disks will be encrypted.  This encryption is a Storage Service Encryption.  To encrypt the disks attached to the VMs, this has to be manually done by following the steps here (https://docs.microsoft.com/en-us/azure/security/azure-security-disk-encryption)