Changed No to Yes for Create Azure Run As Account
My understanding is that for an Azure Automation that is used for runbooks it needs to be a Run As account and that this section is to help you create just such an account. The warnings for if you set it to No make more sense if Yes is what you are being instructed to set it to.