Updated instructions for enabling custom auth
I've added an important note to the the section "How to: Use custom authentication for your application." Without toggling on the App Service Authentication, the environment variable WEBSITE_AUTH_SIGNING_KEY will not be available when hosting the application and authentication will not work. This tidbit of information will save people hours, if not days of research!