Fixing async code fragment
The async example for a C# function does return a `string` where a `Task<string>` is expected. Changing to use `Task.FromResult` though the use of the `async` modifier would be fine too (however, it'd result in a warning of missing any `await` in the body, and I didn't want to add an `await Task.Delay(x)` to the code just to make that go away).