Update partition-data.md
Trying to keep it short. In this document we are explaining the concept and leaving it the code out of this document. Creating some more emphasis on certain things which i see users often forget or overlook.
Made the file shorter. Removed the code from this file, the code should go to a new file and a link be provided for the code.