Supported scenario conflict
We say we support customer custom images (how else is the VHD pre encrypted if not provided from the customer):
"Enable encryption on new IaaS VMs created from pre-encrypted VHD and encryption keys"
but then we say we don't support them:
"Enable encryption on Windows and Linux IaaS VMs customer custom images is NOT supported. Enable enccryption on Linux LVM OS disk is not supported currently. This support will come soon."

Which is it?