Update dailyMessageQuotaUsed metric description
Adding more explanation to the most confusing value - dailyMessageQuotaUsed. Which is a cumulative value holding the total number of messages used since 00:00 UTC up to the point of time being asked. It cannot be split into time grains (each time grain will hold the cumulative daily value, and not the time slice value).