Update iot-hub-devguide-messages-read-custom.md
Added a note about the fact that messages are no longer delivered to the built-in endpoint once you create routes. However, there is an option to add a route to the default built-in endpoint. Something that is not clear in any of the documentation or even stands out in the portal.