Updated incorrect powershell
Updated incorrect powershell that has caused multiple cases as they need to use the New-AzureVnetGateway command instead Net-AzureVirtualNetworkGateway command as it will cause the connection to fail when created as it doesn't create as a primary gateway in the classic Vnet. Since this is for Classic only this affects this.