M12648: Case of ambiguity:  "activities on which email address"
Hello, @sharonlo101,
 
Translator has reported possible source content issue. It seems that "on which email address" means "in which the email address." This proposed file change avoids ambiguity.
 
Please, check this and help to merge if agreed. If not, please, let me know either if I should close this PR, if you would like me to fix it in another way within this PR or if you prefer to fix it in another PR. Could you indicate your PR number in that case so I can check the new version, please?

Many thanks in advance.