M13816: Possible typo mistake: "models"
 Hi, @yzhong94,

Translator has reported possible source content issue. "hardware security models" seems to be a typo of "hardware security modules" because "HSM" is an acronym of "Hardware Security Module." In addition, "Hardware Security Modules" are used in the subsequent segments.

Please, check the proposed file change and help to merge if agreed. 

Many thanks in advance.