M13837: Inconsistent term used
Hello, @thraka,

Translator has reported possible source content issue. As this string is shown in the Imports (an element) section, assumed the string should read "only one imports block" instead of "runtime block."

Please, check the proposed file change and help to merge if agreed. If not, please, let me know either if I should close this PR, if you would like me to fix it in another way within this PR or if you prefer to fix it in another PR. Could you indicate your PR number in that case so I can check the new version, please?

Many thanks in advance.