Application parameters for resource governance
Adding documentation for using application parameters when specifying resource governance. It is important to note that once application parameters are used in Service Fabric version 6.1 or later, it is not possible to downgrade the Service Fabric version. In case when downgrade to version older than 6.1, application manifest validation will fail because it is expecting numerical values for resource governance settings.