Fix for the storageURI output
In it's current format, the storageURI output breaks (Deployment failed. The template output 'storageUri' is not valid: Index (zero based) must be greater than or equal to zero and less than the size of the argument list)

This is fixed once the storageUri type is changed to "object" rather than string. I have also adjusted the value inside the reference function to use the Microsoft.Storage/storageAccounts Resource Provider.