No need for the reader to have GitHub account
The way the quickstart is worded, it does not require the reader to use their own GitHub account, so I think this requirement should be removed from the documentation.

Another option would be to add wording that references the reader's GitHub account...

The _deployment-source-url_ parameter is a sample repository in GitHub that contains a "Hello World" HTTP triggered function. You can replace this URL with your own GitHub repository that contains your function code.