Use case of UDR with Express Route and BGP
In relation to BGP: The original text states "You cannot create user-defined routes if you deploy a virtual network gateway deployed as type: ExpressRoute". This isn't correct. You still can use UDR to force traffic from the ER GW to, for example, an NVA. This is inline with Microsofts reference architectures. It isn't possible to define a User Defined Route to force traffic back to onprem. This is handled by the BGP advertised routes.