Explain that choice of language does not matter
To explain why you can choose either C# or Visual Basic, and why it doesn't matter which you choose. The prior docs ask you to choose one, but gives no context to why / which you should choose. This seems a little strange when following the document in the context of a developer or admin that is not familiar with coding in .Net.