M15127:  Possible type mistake: size
Hello, @iainfoulds, 
Translator has reported possible source content issue. It seems that it should say "side" instead of "size", according to the context. 
 
Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please let me know of your PR number, so we can confirm and close this PR. 

Many thanks in advance.