CC42709: Update-AzureRMVmss incorrect parameter fix #53
Hello, @iainfoulds,

This proposed file change comes from this Community Contribution: https://github.com/MicrosoftDocs/azure-docs.nl-nl/pull/53/files
Please, help to check this proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.