Updated the links to open in new tab
Changed the links for "free account" and Azure portal to open in the new page. Markdown does not have this feature yet, so changed to normal html. The links opening in a new tab let's the reader be on the documentation page and the reader does not need to go back to the documentation if they need to follow the subsequent sections.