"... app settings" can cause confusion
I realize that "Function app settings" are the settings for the function app. However, as you can see from this answer in SO (https://stackoverflow.com/a/47636326/84395), they are being read as the Function's "App Settings" (i.e. Application Settings).

To remove confusion, just name this section based on the name of the tab