M16776:  Text incorrectly locked as a tag: "data fac"
Hello, @spelluru,
Translator has reported source content issue that should be localized. We need to remove the ´´ and replace them by ** for bold type.  

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR. 

Many thanks in advance.