Inaccurate maximum number of data disks for ds2_v2
We have a few incidents related to the accuracy of this document. 
Id 54096227: Unable to attach new disks beyond 32 for Standard_GS4 Virtual Machine. 
Accurate number is 32
Id 53789825: The VM size is ds2_v2 promo which says it supports 8 data disks - maximum number of data disks currently permitted is 4.
Accurate number is 4
I have sent email to related the PM(Meagan McCrory <memccror@microsoft.com> and Jon Beck ).  Meagan confirmed that the numbers are not accurate. But I haven't seen changes been made in this Doc, which blocks me to close the incidents.