Update virtual-networks-name-resolution-for-vms-and-role-instances.md
A recent change started preventing users of Azure App Service using VNET Integration to be able to resolve Azure VM's internal names. Since this wasn't supposed to be allowed since the beginning we need to have it document it so that we can explain to customer what the workaround is.