Correct and simplify file names
The PowerShell file for "IoT hub owner CA" was incorrect, it should be "certs/azure-iot-test-only.root.ca.cert.pem".  If we change this, the filenames between PowerShell and bash become identical (by design).  So... this lets us simplify the filename table to be independent of the script it was generated from.

Could we please incorporate this simplification when you get a chance?  Thanks!